Refactor the test folder
