--- test/evaluation.jl
@@ -1,5 +1,0 @@
-using CounterfactualExplanations.Evaluation
-using DataFrames
-
-bmk = benchmark(counterfactual_data; converge_when=:generator_conditions)
-@test typeof(bmk()) <: DataFrame
