+++ test/generators/probe.jl
@@ -1,19 +1,13 @@
-using Test
-using CounterfactualExplanations
-using CounterfactualExplanations.Data
-using CounterfactualExplanations.Generators
-using CounterfactualExplanations.Objectives
-
 @testset "ProbeGenerator" begin
     @testset "Default arguments" begin
-        generator = ProbeGenerator()
+        generator = Generators.ProbeGenerator()
         @test typeof(generator) <: AbstractGenerator
         @test generator.λ == 0.1
         @test generator.loss == Flux.Losses.logitbinarycrossentropy
     end
 
     @testset "Custom arguments" begin
-        generator = ProbeGenerator(; λ=0.5, loss=:mse)
+        generator = Generators.ProbeGenerator(; λ=0.5, loss=:mse)
         @test generator.λ == 0.5
         @test generator.loss == Flux.Losses.mse
     end
@@ -21,15 +15,15 @@ end
 
 @testset "hinge_loss" begin
     @testset "Hinge loss calculation" begin
-        counterfactual_data = load_linearly_separable()
-        M = fit_model(counterfactual_data, :Linear)
+        counterfactual_data = Data.load_linearly_separable()
+        M = Models.fit_model(counterfactual_data, :Linear)
         target = 2
         factual = 1
-        chosen = rand(findall(predict_label(M, counterfactual_data) .== factual))
-        x = select_factual(counterfactual_data, chosen)
+        chosen = rand(findall(Models.predict_label(M, counterfactual_data) .== factual))
+        x = DataPreprocessing.select_factual(counterfactual_data, chosen)
         # Search:
-        generator = ProbeGenerator()
-        linear_counterfactual = generate_counterfactual(
+        generator = Generators.ProbeGenerator()
+        linear_counterfactual = CounterfactualExplanations.generate_counterfactual(
             x,
             target,
             counterfactual_data,
@@ -40,8 +34,8 @@ end
             invalidation_rate=0.1,
             learning_rate=0.1,
         )
-        loss = hinge_loss(linear_counterfactual)
-        rate = invalidation_rate(linear_counterfactual)
+        loss = Generators.hinge_loss(linear_counterfactual)
+        rate = Generators.invalidation_rate(linear_counterfactual)
         @test rate <= 0.1
         @test loss <= 0.9
     end
