+++ test/models/generative_models.jl
@@ -1,5 +1,3 @@
-using MLUtils
-
 # Setup
 M = synthetic[:classification_binary][:models][:MLP][:model]
 counterfactual_data = synthetic[:classification_binary][:data]
@@ -8,10 +6,10 @@ ys = counterfactual_data.y
 generator = generator_catalogue[:revise]()
 
 # Counterfactual search
-x = select_factual(counterfactual_data, rand(1:size(X, 2)))
-y = predict_label(M, counterfactual_data, x)
+x = DataPreprocessing.select_factual(counterfactual_data, Random.rand(1:size(X, 2)))
+y = Models.predict_label(M, counterfactual_data, x)
 target = get_target(counterfactual_data, y[1])
-ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
+ce = CounterfactualExplanations.generate_counterfactual(x, target, counterfactual_data, M, generator)
 
 using CounterfactualExplanations.GenerativeModels: retrain!
 CounterfactualExplanations.GenerativeModels.retrain!(
