+++ test/other/evaluation.jl
@@ -1,0 +1,2 @@
+bmk = Evaluation.benchmark(counterfactual_data; converge_when=:generator_conditions)
+@test typeof(bmk()) <: DataFrame
